move JobDetails page's components to reusable folder
